Fixed tenant grpc file
